requirements.txt
requests
selenium 
pyyaml
2captcha-python
undetected-chromedriver
webdriver-manager
setuptools